[issue-3] move vector-specific functions to Vector
